requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements-dev.txt requirements-dev.in
#
appnope==0.1.2
    # via ipython
asgiref==3.3.1
    # via
    #   -r requirements.txt
    #   django
attrs==20.3.0
    # via
    #   -r requirements.txt
    #   flake8-bugbear
    #   jsonschema
    #   pytest
backcall==0.2.0
    # via ipython
certifi==2020.12.5
    # via
    #   -r requirements.txt
    #   requests
chardet==4.0.0
    # via
    #   -r requirements.txt
    #   requests
coreapi==2.3.3
    # via
    #   -r requirements.txt
    #   drf-yasg
coreschema==0.0.4
    # via
    #   -r requirements.txt
    #   coreapi
    #   drf-yasg
coverage==5.3.1
    # via pytest-cov
decorator==4.4.2
    # via ipython
django-environ==0.4.5
    # via -r requirements.txt
django-extensions==3.1.0
    # via -r requirements.txt
django-filter==2.4.0
    # via -r requirements.txt
django==3.1.5
    # via
    #   -r requirements.txt
    #   django-filter
    #   djangorestframework
    #   drf-yasg
djangorestframework==3.12.2
    # via
    #   -r requirements.txt
    #   drf-yasg
drf-yasg[validation]==1.20.0
    # via -r requirements.txt
factory-boy==3.2.0
    # via
    #   -r requirements-dev.in
    #   pytest-factoryboy
faker==5.4.0
    # via factory-boy
flake8-bugbear==20.11.1
    # via -r requirements-dev.in
flake8==3.8.4
    # via
    #   -r requirements-dev.in
    #   flake8-bugbear
idna==2.10
    # via
    #   -r requirements.txt
    #   requests
inflection==0.5.1
    # via
    #   -r requirements.txt
    #   drf-yasg
    #   pytest-factoryboy
iniconfig==1.1.1
    # via pytest
ipdb==0.13.4
    # via -r requirements-dev.in
ipython-genutils==0.2.0
    # via traitlets
ipython==7.19.0
    # via
    #   -r requirements-dev.in
    #   ipdb
isort==5.7.0
    # via -r requirements-dev.in
itypes==1.2.0
    # via
    #   -r requirements.txt
    #   coreapi
jedi==0.18.0
    # via ipython
jinja2==2.11.2
    # via
    #   -r requirements.txt
    #   coreschema
jsonschema==3.2.0
    # via
    #   -r requirements.txt
    #   swagger-spec-validator
markupsafe==1.1.1
    # via
    #   -r requirements.txt
    #   jinja2
mccabe==0.6.1
    # via flake8
packaging==20.8
    # via
    #   -r requirements.txt
    #   drf-yasg
    #   pytest
parso==0.8.1
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pluggy==0.13.1
    # via pytest
prompt-toolkit==3.0.10
    # via ipython
psycopg2-binary==2.8.6
    # via -r requirements.txt
ptyprocess==0.7.0
    # via pexpect
py==1.10.0
    # via pytest
pycodestyle==2.6.0
    # via flake8
pyflakes==2.2.0
    # via flake8
pygments==2.7.3
    # via ipython
pyparsing==2.4.7
    # via
    #   -r requirements.txt
    #   packaging
pyrsistent==0.17.3
    # via
    #   -r requirements.txt
    #   jsonschema
pytest-cov==2.10.1
    # via -r requirements-dev.in
pytest-django==4.1.0
    # via -r requirements-dev.in
pytest-factoryboy==2.1.0
    # via -r requirements-dev.in
pytest==6.2.1
    # via
    #   -r requirements-dev.in
    #   pytest-cov
    #   pytest-django
    #   pytest-factoryboy
python-dateutil==2.8.1
    # via faker
pytz==2020.5
    # via
    #   -r requirements.txt
    #   django
pyyaml==5.3.1
    # via
    #   -r requirements.txt
    #   swagger-spec-validator
requests==2.25.1
    # via
    #   -r requirements.txt
    #   coreapi
ruamel.yaml.clib==0.2.2
    # via
    #   -r requirements.txt
    #   ruamel.yaml
ruamel.yaml==0.16.12
    # via
    #   -r requirements.txt
    #   drf-yasg
six==1.15.0
    # via
    #   -r requirements.txt
    #   jsonschema
    #   python-dateutil
    #   swagger-spec-validator
sqlparse==0.4.1
    # via
    #   -r requirements.txt
    #   django
swagger-spec-validator==2.7.3
    # via
    #   -r requirements.txt
    #   drf-yasg
text-unidecode==1.3
    # via faker
toml==0.10.2
    # via pytest
traitlets==5.0.5
    # via ipython
uritemplate==3.0.1
    # via
    #   -r requirements.txt
    #   coreapi
    #   drf-yasg
urllib3==1.26.2
    # via
    #   -r requirements.txt
    #   requests
wcwidth==0.2.5
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
